Modify grammar to new conditional GOTO
After review, I find the use of this grammar rule problematic:

stmt:
  IF exp THEN stmt

In this way stmt is recursively defined and an instruction object
would need a pointer to yet another instruction object, making
instruction becoming an abstract syntax tree node. This is the
complexity I have been trying to avoid from the beginning.

Should stmt in IF be replaced, merging IF with GOTO seems to be
an easier way out.

As for the new production rule:

stmt:
  GOTO LABEL IF REG  { ... }

REG part could not be exp construct. When instruction object is
constructed, exp is captured, making the conditional always eval
to the same value. Consider below stmt:

goto L10 if #rb

#rb is evaluated to integer value and captured by instruction
object. Intended effect of IF conditional is nullified.